Drop NSDateFormatter cache attached to the Thread (NSDateFormatter is now thread safe). Allow a custom date formatter to be set on the logger.
THREAD SAFETY

On iOS 7 and later NSDateFormatter is thread safe.

On OS X v10.9 and later NSDateFormatter is thread safe so long as you are using the modern behavior in a 64-bit app.

On earlier versions of the operating system, or when using the legacy formatter behavior or running in 32-bit on OS X, NSDateFormatter is not thread safe and you therefore must not mutate a date formatter simultaneously from multiple threads.