Fixed up the existing Xcode project file.
This includes:

• Moving it up a level in the folder hierarchy, so that it’s no longer
in ‘Sources’, but rather at the package root.
• Correcting the paths to all the various files within the project,
both to account for the move of the project file itself as well as the
other moves made in previous commits to support the Swift Package
Manager.
• Renaming a few things (like XGCLogger.h to ObjcExceptionBridging.h)
to satisfy Xcode’s assumptions about file names (Swift Package Manager
blessedly doesn’t care).
• Happened to fix the target & scheme naming from “OS X” to “macOS”.